Introduce user files count
This part is not that easy. When bp_is_my_profile() it is ok, but as files can be private, the real count when the user displayed is not the logged in user should not be displayed. Instead it should be the public count.
By default every file has the publish status, privacy is stored in post meta, a future enhancement would be to use the private status and the password feature of WordPress post.
I will work on this for 1.3. Untill this version, i am using a javascript workaround to display the user count and introduce a new function buddydrive_count_user_files().
Using this function with some filters is fixing #7.
The javascript workaround will be a first reply to #8.