Fix:  Allow BuddyDrive to send files larger than the WordPress Memory Limit
If your WP memory limit is 256M as an example, and PHP's output
buffering is being used anywhere on the site, and you are downloading a
file > 256M (or whatever the WP memory limit is), output buffering must
be turned off before PHP's readfile occurs.  Now, files can be of any
size regardless of ob's level, and regardless of WP's memory limit.