Include the french translation of the plugin.
WP Idea Stream made me realize, as a plugin author, the only way to be sure to avoid regressions about translations is to include the files. So i will now allways include my version of the translation for the french language so that french people will not have to wait for the busy french translators to find the time to translate the plugin.